add encoding of tagged values
I'm not sure if this library should attempt to provide first-class
support for conversion to/from whatever native types might be
appropriate for the various tag meanings (dates, bignums, etc). It would
add a lot of dependencies and still might not match exactly what someone
would like. The CBOR RFC doesn't always specify exactly what the meaning
of the various tags are, so I think for now it's best left the caller to
figure out how they want to interpret the tags.